+++ prometheus_fastapi_exporter/__init__.py
@@ -1,3 +1,0 @@
-from . import exporter
-
-PrometheusFastApiExporter = exporter.PrometheusFastApiExporter